Avoid failing `RunJosmTask` when the current plugin does not require other plugins
The problem was, that the file `build/josm-custom-config/requiredPlugins.xml` was not generated. Because that file was required by the tasks of type `RunJosmTask`, these tasks failed.
The generation process of that file is now done in the task `writePluginConfig`, while the copying of the plugin *.jar files is now done by the task `updateJosmPlugins` instead of both `updateCurrentPlugin` and `updatePluginDependencies` (which both no longer exist).