Implement getAllRequiredJosmPlugins() and getNextJosmVersion() in Kotlin and use Gradle's built-in dependency resolution
Both methods are now implemented as extensions for the class org.gradle.api.Project .
They use detached configurations to resolve in one case all required plugins (including transitive dependencies) given a list of only the directly required JOSM plugins. And in the other case starting with a version number the next available JOSM version is determined.
The dependencies which are found, are returned as org.gradle.api.artifacts.Dependency objects for convenience.