Split the temporary JOSM directories into three places, one for preferences, cache and user data
The `cleanJosm` task now has three tasks on which it depends, one for each directory.
The `runJosm` and `debugJosm` tasks pick up the three directories as set in the configuration. By default they are `build/.josm/cache`, `build/.josm/pref` and `build/.josm/userdata`.
If you want to use a (very old) JOSM version before 7841, the `build/.josm/pref` directory is used as `josm.home` with everything in it (cache in a subdirectory).