Download JOSM plugin list as Gradle dependency
The list is downloaded to find so called virtual plugins, that represent groups of native plugins for OSX, Windows and Unixoid OSes. Previously it was read directly, now it will be processed through Gradles dependency mechanism.
This makes sure, the list is cached and e.g. available even when offline (if the list has previously been downloaded when online).
This fixes https://gitlab.com/floscher/gradle-josm-plugin/issues/8 .